git: cvsexportcommit
